+++ include/sentry.h
@@ -30,7 +30,7 @@ extern "C" {
 #        define SENTRY_SDK_NAME "sentry.native"
 #    endif
 #endif
-#define SENTRY_SDK_VERSION "0.7.20"
+#define SENTRY_SDK_VERSION "0.8.1"
 #define SENTRY_SDK_USER_AGENT SENTRY_SDK_NAME "/" SENTRY_SDK_VERSION
 
 /* common platform detection */
@@ -1416,25 +1416,38 @@ SENTRY_API sentry_uuid_t sentry_capture_event(sentry_value_t event);
  * Allows capturing independently created minidumps.
  *
  * This generates a fatal error event, includes the scope and attachments.
- * If the event isn't dropped by a before-send hook, the minidump is attached
+ * If a before-send hook doesn't drop the event, the minidump is attached
  * and the event is sent.
  *
  * Returns a nil `UUID` if capturing the minidump failed and the event-id
  * otherwise. Uploads can fail because capturing is asynchronous, so a non-nil
  * `UUID` is not a delivery guarantee. However, if the minidump is successfully
  * delivered, the ID is guaranteed to be the same as the event in the Sentry UI.
+ *
+ * Note: You don't need this function if you rely on Sentry to create the
+ * minidump. This is useful when you have a minidump captured through a
+ * different mechanism, and you want Sentry to ingest it.
  */
 SENTRY_API sentry_uuid_t sentry_capture_minidump(const char *path);
 SENTRY_API sentry_uuid_t sentry_capture_minidump_n(
     const char *path, size_t path_len);
 
 /**
- * Captures an exception to be handled by the backend.
+ * Captures a system-native exception that you retrieve when you manually handle
+ * `POSIX` signals or `SEH` exceptions and want to keep using that handling
+ * instead of the top-level handlers in our backends. The exception is still
+ * processed as a sentry event inside the SDK, including applying scope metadata
+ * and invoking hooks.
+ *
+ * The passed in `sentry_ucontext_t` must be filled with the OS-specific
+ * exception data (as specified in the struct definition) that you retrieve
+ * from your handler.
  *
  * This is safe to be called from a crashing thread and may not return.
  *
  * Note: The `crashpad` client currently supports this only on Windows. `inproc`
- *       and `breakpad` support it on all platforms.
+ *       and `breakpad` supports it on all platforms (on macOS, the `uctx`
+ *       argument is ignored when using the `breakpad` backend).
  */
 SENTRY_EXPERIMENTAL_API void sentry_handle_exception(
     const sentry_ucontext_t *uctx);
@@ -1507,6 +1520,15 @@ SENTRY_API void sentry_set_fingerprint_n(
  * Removes the fingerprint.
  */
 SENTRY_API void sentry_remove_fingerprint(void);
+
+/**
+ * Set the trace. The primary use for this is to allow other SDKs to propagate
+ * their trace context to connect events on all layers
+ */
+SENTRY_API void sentry_set_trace(
+    const char *trace_id, const char *parent_span_id);
+SENTRY_API void sentry_set_trace_n(const char *trace_id, size_t trace_id_len,
+    const char *parent_span_id, size_t parent_span_id_len);
 
 /**
  * Sets the transaction.